+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/case-view/case-view.model.ts
@@ -41,18 +41,3 @@ export class CaseView {
   };
   public case_flag?: Flags;
 }
-
-export interface ServiceOrg {
-  business_are: string;
-  ccd_case_types: string;
-  ccd_service_name: string;
-  jurisdiction: string;
-  last_update?: string;
-  org_unit: string;
-  service_code: string;
-  service_description: string;
-  service_id: number;
-  service_short_description: string;
-  sub_business_area: string;
-
-}
