+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/case-view/service-org-response.model.ts
@@ -1,0 +1,13 @@
+export interface ServiceOrg {
+    business_are: string;
+    ccd_case_types: string;
+    ccd_service_name: string;
+    jurisdiction: string;
+    last_update?: string;
+    org_unit: string;
+    service_code: string;
+    service_description: string;
+    service_id: number;
+    service_short_description: string;
+    sub_business_area: string;
+  }
